youtube_rest_api/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
python-decouple==3.5
pytz==2021.3
pyzmq==22.1.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
wincertstore==0.2
